asfdemux/asfpacket: specify the offset by an amount of filled buffer size when concatenating fragmented media objects
This change is preliminary for inserting VC-1 startcodes of advanced
profile. To achieve this, the VC-1 startcode size is required to be
added to the amount of filled buffer size of a payload. As a result
the offset to concatenate fragmented media objects couldn't be determined
by mo_offset of an asfpacket, so the offset is specified as the amount
of filled buffer size instead, which will take account into the VC-1
start code size.